requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118

torch==2.0.1+cu118
opencv-python
onnxruntime-gpu
insightface
imageio
pillow